+++ messages/fi/config.php
@@ -1,12 +1,11 @@
 <?php
-
-return [
-    '<strong>Task</strong> module configuration' => '',
-    'Max tasks items' => '',
-    'Show snippet' => '',
-    'Show snippet in Space' => '',
-    'Shows a widget with tasks on the dashboard where you are assigned/responsible.' => '',
-    'Shows the widget also on the dashboard of spaces.' => '',
-    'Sort order' => '',
-    'Your tasks snippet' => '',
-];
+return array (
+  '<strong>Task</strong> module configuration' => '<strong>Tehtävät</strong> laajennuksen hallinta',
+  'Max tasks items' => 'Maksimi tehtävien määrä',
+  'Show snippet' => 'Näytä kappale',
+  'Show snippet in Space' => 'Näytä kappalle sivulla',
+  'Shows a widget with tasks on the dashboard where you are assigned/responsible.' => 'Näyttää widgetin, joka näyttää tähtävät joihin sinut on määrätty/vastuussa.',
+  'Shows the widget also on the dashboard of spaces.' => 'Näyttää widgetti myös sivujen etusivuilla.',
+  'Sort order' => 'Lajittelujärjestys',
+  'Your tasks snippet' => 'Tehtäväsi kappale',
+);