+++ messages/fi/views_activities_TaskFinished.php
@@ -1,5 +1,4 @@
 <?php
-
-return [
-    '{userName} finished task {task}.' => '',
-];
+return array (
+  '{userName} finished task {task}.' => '{userName} sai tehtävän {task} valmiiksi.',
+);