+++ messages/fr/models_forms_TaskFilter.php
@@ -1,9 +1,15 @@
 <?php
-return array (
-  'Created by me' => 'Créé par moi',
-  'Filter tasks' => 'Filtrer les tâches',
-  'I\'m assigned' => 'Je suis assigné',
-  'I\'m responsible' => 'Je suis responsable',
-  'Overdue' => 'Échue',
-  'Status' => 'Statut',
-);
+
+return [
+    'End date' => '',
+    'Filter status' => '',
+    'Spaces' => '',
+    'Start date' => '',
+    'Title' => '',
+    'Created by me' => 'Créé par moi',
+    'Filter tasks' => 'Filtrer les tâches',
+    'I\'m assigned' => 'Je suis assigné',
+    'I\'m responsible' => 'Je suis responsable',
+    'Overdue' => 'Échue',
+    'Status' => 'Statut',
+];