+++ messages/hu/models_forms_TaskFilter.php
@@ -1,9 +1,15 @@
 <?php
-return array (
-  'Created by me' => 'Általam létrehozva',
-  'Filter tasks' => 'Feladatok szűrése',
-  'I\'m assigned' => 'Én vagyok kijelölve a feladatra',
-  'I\'m responsible' => 'Én vagyok a felelőse',
-  'Overdue' => 'Lejárt a határidő',
-  'Status' => 'Állapot',
-);
+
+return [
+    'End date' => '',
+    'Filter status' => '',
+    'Spaces' => '',
+    'Start date' => '',
+    'Title' => '',
+    'Created by me' => 'Általam létrehozva',
+    'Filter tasks' => 'Feladatok szűrése',
+    'I\'m assigned' => 'Én vagyok kijelölve a feladatra',
+    'I\'m responsible' => 'Én vagyok a felelőse',
+    'Overdue' => 'Lejárt a határidő',
+    'Status' => 'Állapot',
+];