+++ messages/nl/models_forms_TaskFilter.php
@@ -1,9 +1,15 @@
 <?php
-return array (
-  'Created by me' => 'Gemaakt door mij',
-  'Filter tasks' => 'Filter taken',
-  'I\'m assigned' => 'Toegewezen aan mij',
-  'I\'m responsible' => 'Mijn verantwoordelijkheid',
-  'Overdue' => 'Over tijd',
-  'Status' => 'Voortgang',
-);
+
+return [
+    'End date' => '',
+    'Filter status' => '',
+    'Spaces' => '',
+    'Start date' => '',
+    'Title' => '',
+    'Created by me' => 'Gemaakt door mij',
+    'Filter tasks' => 'Filter taken',
+    'I\'m assigned' => 'Toegewezen aan mij',
+    'I\'m responsible' => 'Mijn verantwoordelijkheid',
+    'Overdue' => 'Over tijd',
+    'Status' => 'Voortgang',
+];