+++ docs/CHANGELOG.md
@@ -3,12 +3,16 @@ Changelog
 
 1.2.3
 --------------------
-- Enh: Filter by cotnainer related tasks in "Your tasks" snippet.
+- Enh: Filter by container related tasks in "Your tasks" snippet.
 - Enh: Global overview
 - Enh: Added move content feature
 - Enh: Added Tabbed Menu to single task view
 - Enh: Added `searchPaginationSize` module configuration 
 - Enh: Added `showTopMenuItem` and `topMenuSort` module configuration
+- Fix: RemindEnd notification displays start date instead of end date
+- Enh: Added date filter to search view
+- Enh: Enhanced list view usability
+- Enh: Use of new richtext editor
 
 1.2.2 (October 15, 2018)
 --------------------